Just the sum of Args.
Create a class named SumOfArgs that will print out the sum of all
integer arguments found on the command line. It should have a main()
method so the class can be run. Anything that is not an integer should
be skipped (should not contribute to the sum). It should print out only
the sum (nothing else!). No error messages can be printed out (no
matter what the command line args look like).