Substring
Create a class named Substring that will expect the first command line
argument to be a string, and the second two command line arguments to
be integers, the first will be used as an index and the second as a
length. The output should be the subtring of string starting at the
index and of the specified length.